pyproject.toml
[build-system]
build-backend = "maturin"
requires = ["maturin>=1.6,<2.0"]

[project]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Programming Language :: Rust",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Distributed Computing",
]
dependencies = [
    "opentelemetry-api>=1.26.0,<2.0.0",
    "threadsafe-async>=0.2.1,<1.0.0",
]
description = "High-performance Python bindings for a Rust-based Kafka client"
dynamic = ["version"]
keywords = ["async", "client", "distributed", "kafka", "messaging"]
name = "prosody"
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "opentelemetry-sdk>=1.26.0,<2.0.0",
    "opentelemetry-exporter-otlp>=1.26.0,<2.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-watch>=4.2.0",
    "pytest>=7.3.1",
]

[tool.maturin]
features = ["pyo3/extension-module"]
python-source = "python"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
